shadow: more resilient and correct shadow hook
The previous implementation from v2 runs after compiling the cljs files,
since that's the only point in the shadow hooks where the list of
recompiled files is available. This is problematic though. Imagine
adding

  #?(:clj (def foo 1))

and using it in an `e/server` block. Since the cljs compiler runs first
the namespace may not be reloaded on clj and the compiler will fail,
throwing an exception.

The new implementation monkey-patches the shadow compiler in order to
recompile on clj first, fixing the use case above.